Organize project structure
Add documentation on public classes
Add copyright headers
