+++ StaffMod/src/main/kotlin/opekope2/avm_staff/api/item/StaffItem.kt
@@ -1,0 +1,11 @@
+// Copyright (c) 2023-2024 opekope2
+// Staff Mod is licensed under the MIT license: https://github.com/opekope2/StaffMod/blob/main/LICENSE
+
+package opekope2.avm_staff.api.item
+
+import net.minecraft.item.Item
+
+/**
+ * Staff item.
+ */
+class StaffItem(settings: Settings) : Item(settings)
