+++ StaffMod/src/main/kotlin/opekope2/avm_staff/internal/StaffMod.kt
@@ -1,4 +1,7 @@
-package opekope2.avm_staff
+// Copyright (c) 2023-2024 opekope2
+// Staff Mod is licensed under the MIT license: https://github.com/opekope2/StaffMod/blob/main/LICENSE
+
+package opekope2.avm_staff.internal
 
 import net.fabricmc.api.ModInitializer
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings
@@ -8,15 +11,16 @@ import net.minecraft.item.Items
 import net.minecraft.registry.Registries
 import net.minecraft.registry.Registry
 import net.minecraft.util.Identifier
-import opekope2.avm_staff.item.StaffItem
-import opekope2.avm_staff.packet.c2s.AddBlockToStaffC2SPacket
-import opekope2.avm_staff.packet.c2s.RemoveBlockFromStaffC2SPacket
-import opekope2.avm_staff.server.ServerStaffHandler
+import opekope2.avm_staff.api.item.StaffItem
+import opekope2.avm_staff.internal.packet.c2s.AddBlockToStaffC2SPacket
+import opekope2.avm_staff.internal.packet.c2s.RemoveBlockFromStaffC2SPacket
+import opekope2.avm_staff.internal.server.StaffPacketHandler
 
 @Suppress("unused")
 object StaffMod : ModInitializer {
     const val MOD_ID = "avm_staff"
 
+    @JvmField
     val STAFF_ITEM: StaffItem = Registry.register(
         Registries.ITEM,
         Identifier(MOD_ID, "staff"),
@@ -30,7 +34,7 @@ object StaffMod : ModInitializer {
         ItemGroupEvents.modifyEntriesEvent(ItemGroups.COMBAT).register { entries ->
             entries.addAfter(Items.TRIDENT, STAFF_ITEM)
         }
-        AddBlockToStaffC2SPacket.registerGlobalReceiver(ServerStaffHandler::addBlockToStaff)
-        RemoveBlockFromStaffC2SPacket.registerGlobalReceiver(ServerStaffHandler::removeBlockFromStaff)
+        AddBlockToStaffC2SPacket.registerGlobalReceiver(StaffPacketHandler::addBlockToStaff)
+        RemoveBlockFromStaffC2SPacket.registerGlobalReceiver(StaffPacketHandler::removeBlockFromStaff)
     }
 }
