+++ StaffMod/src/main/kotlin/opekope2/avm_staff/internal/model/StaffItemModel.kt
@@ -1,4 +1,7 @@
-package opekope2.avm_staff.model
+// Copyright (c) 2023-2024 opekope2
+// Staff Mod is licensed under the MIT license: https://github.com/opekope2/StaffMod/blob/main/LICENSE
+
+package opekope2.avm_staff.internal.model
 
 import net.fabricmc.api.EnvType
 import net.fabricmc.api.Environment
@@ -16,8 +19,8 @@ import net.minecraft.item.BlockItem
 import net.minecraft.item.ItemStack
 import net.minecraft.util.math.Direction
 import net.minecraft.util.math.random.Random
-import opekope2.avm_staff.util.staffHasItem
-import opekope2.avm_staff.util.staffItem
+import opekope2.avm_staff.util.isItemInStaff
+import opekope2.avm_staff.util.itemInStaff
 import org.joml.Vector3f
 import java.util.function.Supplier
 
@@ -30,9 +33,9 @@ class StaffItemModel(model: BakedModel) : ForwardingBakedModel() {
     override fun emitItemQuads(stack: ItemStack, randomSupplier: Supplier<Random>, context: RenderContext) {
         super.emitItemQuads(stack, randomSupplier, context)
 
-        if (!stack.staffHasItem) return
+        if (!stack.isItemInStaff) return
 
-        val itemStack = stack.staffItem
+        val itemStack = stack.itemInStaff
         val item = itemStack?.item as? BlockItem
         if (item == null) {
             // TODO render item model
