--- StaffMod/src/main/kotlin/opekope2/avm_staff/item/StaffItem.kt
@@ -1,5 +1,0 @@
-package opekope2.avm_staff.item
-
-import net.minecraft.item.Item
-
-class StaffItem(settings: Settings) : Item(settings)
